Break out type definitions into separate files
