+++ AnsibleTowerClasses/AnsibleTower/IAnsibleObject.cs
@@ -1,0 +1,16 @@
+using Newtonsoft.Json;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using Newtonsoft.Json.Linq;
+
+namespace AnsibleTower
+{
+    public interface IAnsibleObject {
+        int id { get; set; }
+        string type { get; set; }
+        string url { get; set; }
+        DateTime created { get; set; }
+        Tower AnsibleTower { get; set; }
+    }
+}
