+++ AnsibleTowerClasses/AnsibleTower/Organization.cs
@@ -1,0 +1,22 @@
+using Newtonsoft.Json;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using Newtonsoft.Json.Linq;
+
+namespace AnsibleTower
+{
+    public class Organization : IAnsibleObject
+    {
+        public int id { get; set; }
+        public string type { get; set; }
+        public string url { get; set; }
+        //public Related related { get; set; }
+        //public SummaryFields summary_fields { get; set; }
+        public DateTime created { get; set; }
+        public DateTime modified { get; set; }
+        public string name { get; set; }
+        public string description { get; set; }
+        public Tower AnsibleTower { get; set; }
+    }
+}
